requirements.txt
Click==7.0
Flask==1.0.2
itsdangerous==1.1.0
Jinja2==2.10
MarkupSafe==1.1.0
PyYAML==3.13
twitter==1.18.0
Werkzeug==0.14.1
